Update token check
